refactor(dsl-services-dotnet-asp): collapse route builder duplication
